+++ apps/wow-docs/components/ImageCards.tsx
@@ -3,6 +3,7 @@ import Space from "@components/Space";
 import Title from "@components/Text/Title";
 import Image from "next/image";
 import type { CSSProperties, ReactNode } from "react";
+import { Fragment } from "react";
 
 export type Item = {
   main: string;
@@ -21,7 +22,7 @@ interface ImageCardProps {
 const ImageCards = ({ items }: ImageCardProps) => {
   return items.map(
     ({ main, sub, imageAlt, imageSrc, imageWidth, imageHeight, cardStyle }) => (
-      <>
+      <Fragment key={imageSrc}>
         <Title main={main} sub={sub} variant="component" />
         <Space height={20} />
         <Card style={cardStyle}>
@@ -33,7 +34,7 @@ const ImageCards = ({ items }: ImageCardProps) => {
           />
         </Card>
         <Space height={40} />
-      </>
+      </Fragment>
     )
   );
 };