+++ apps/wow-docs/constants/routePath.ts
@@ -34,7 +34,7 @@ export const componentPaths = {
   stepper: "stepper",
   switch: "switch",
   table: "table",
-  tabs: "tabs",
+  tabBar: "tabbar",
   tag: "tag",
   textButton: "textbutton",
   textField: "textfield",
@@ -70,7 +70,7 @@ type RoutePath = {
     stepper: string;
     switch: string;
     table: string;
-    tabs: string;
+    tabBar: string;
     tag: string;
     textButton: string;
     textField: string;