+++ apps/wow-docs/constants/store/searchBarData.ts
@@ -1,0 +1,20 @@
+import type { Item } from "@components/ImageCards";
+
+export const widthSearchBarItems: Item[] = [
+  {
+    main: "PC",
+    sub: "그리드 너비에 맞게 고정하여 사용",
+    imageAlt: "searchbar-pc",
+    imageSrc: "/component/searchbar/searchbar-pc.svg",
+    imageWidth: 907,
+    imageHeight: 148,
+  },
+  {
+    main: "Mobile",
+    sub: "전체 그리드 혹은 박스 컴포넌트에 Fill로 너비 맞추기",
+    imageAlt: "searchbar-mobile",
+    imageSrc: "/component/searchbar/searchbar-mobile.svg",
+    imageWidth: 390,
+    imageHeight: 254,
+  },
+];